+++ app/Http/Controllers/Admin/Tag/DeleteController.php
@@ -9,7 +9,11 @@ class DeleteController extends Controller
 {
     public function __invoke(Tag $tag)
     {
-        $tag->delete();
+        if (auth()->user()->role == 0) {
+            $tag->delete();
+        }else{
+            abort(504);
+        }
         return redirect()->route('admin.tag.index');
     }
 }
